[ALPS09758296] Fix coverity printf specifiers
[Description]
ALPS09758296, ALPS09857016, ALPS09014548

[Test]
Build pass

MTK-Commit-Id: a0fe047242851de766cd0056ebe7cdee8fa3ac0a

CR-Id: ALPS09758296
Feature: Wi-Fi NAN
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/10421896
Commit-Check: srv_check_service <srv_check_service@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
AutoUT-Review-Label: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Build: srv_preflight_b001 <srv_preflight_b001@mediatek.com>
Test: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Coverity-Review-Label: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Test: srv_pf_nep_sanity
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
ODB-Check: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/10451483
GAI-Code-Review: srv_ai_review001 <srv_ai_review001@mediatek.com>

+++ nan/nanScheduler.c
@@ -4835,8 +4835,8 @@ uint32_t nanGetCommonBandWithConcurrent(struct ADAPTER *prAdapter,
 			       rP2pChnlInfo.u4PrimaryChnl,
 			       u4CommonSupportedBand,
 			       u4CommonSupportedBand &
-				~(BIT(ENUM_SUPPORTED_BN_6G) |
-				  BIT(ENUM_SUPPORTED_BN_5G_LOW)));
+			       (uint32_t)(~(BIT(ENUM_SUPPORTED_BN_6G) |
+					    BIT(ENUM_SUPPORTED_BN_5G_LOW))));
 			u4CommonSupportedBand &=
 				~(BIT(ENUM_SUPPORTED_BN_6G) |
 				  BIT(ENUM_SUPPORTED_BN_5G_LOW));
@@ -13785,7 +13785,7 @@ nanSchedCmdUpdateCRB(struct ADAPTER *prAdapter, uint32_t u4SchIdx)
 			continue;
 
 		DBGLOG(NAN, INFO,
-		       "sch:%u, %u, MapId=%u, avail=%02x-%02x-%02x-%02x\n",
+		       "sch:%u, %zu, MapId=%u, avail=%02x-%02x-%02x-%02x\n",
 		       u4SchIdx, i, prTimeline->ucMapId,
 		       ((uint8_t *)(prTimeline->au4AvailMap))[0],
 		       ((uint8_t *)(prTimeline->au4AvailMap))[1],
